refactor: extract BackgroundCoordinator and worker module, modularize UI components, and implement OCR caching
