+++ src/core/mod.rs
@@ -2,4 +2,6 @@ pub mod model;
 pub mod ports;
 pub mod types;
 pub mod usecases;
+pub mod worker;
+pub mod coordinator;
