+++ src/infra/mod.rs
@@ -1,2 +1,3 @@
 pub mod settings;
+pub mod win32;
