+++ src/ui/mod.rs
@@ -1,6 +1,7 @@
 mod app;
 mod crop_overlay;
 mod popups;
+pub mod components;
 
 pub use app::App;
